Fix cyclic dependency between iallocator and pools
The IP pool management infrastructure currently needs to know an instance's
node to fill in IP addresses. For this to happen, the iallocator must run
before we generate IP addresses, yet the iallocator itself needs the instance's
NIC dictionaries.

We break this cycle by passing ip="pool" to the iallocator and filling-in the
IP address after the iallocator run.

Signed-off-by: Apollon Oikonomopoulos <apollon@noc.grnet.gr>