Hack bdev.py to monitor drbd removal
This patch hacks the DRBD remove code and runs an external script
before the `drbdsetup down' command. The script logs the current
state to a separate file to help find out who uses the drbd device.
The script path is hardcoded inside the code and should be:

/etc/ganeti/debug-drbd-remove.sh

Make sure the script has execute permissions set.

!THIS COMMIT SHOULD BE REVERTED AFTER WE FIND THE PROBLEM!

Signed-off-by: Constantinos Venetsanopoulos <cven@grnet.gr>