rapi: Pass existing depends body arg to opcode
Most rlib2 classes override default _GetDefaultData() method with
custom methods that parse request's body and query args and return a
(body, specific_static) tuple eventually passed to FillOpCode().
Job dependencies are defined in depends body argument that might get
"lost" because most of those methods return {} as body.  So in order
not to modify every custom method we update the returned body with
the depends argument of original request body for all rapi calls
inside OpcodeResource._GenericHandler(). So it's up to the client
to include `depends` argument including the job dependencies if desired.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>