Do not abort if hotplug is not supported
This will be useful for an external entity using RAPI that
wants to do device modifications to instances.

The common use case for that is:
"I want to add a nic/disk to an instance. If it is running
then hotplug the device. If not, then just add it to config."

Another use case is that if we are unaware of the instance's
hypervisor: "I want do add a nic/disk to an instance and if
its hypervisor is KVM try to hotplug it."

For all those cases just print a warning in CheckPrereq and
disable any further hotplug related actions (and let modifications
take place after reboot).

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>