Fix NODE/NODE_RES locking in LUInstanceCreate
Both NODE and NODE_RES locks were acquired opportunistically if so
requested by the user. LUInstanceCreate requires, however, that the
actually locked elements on NODE and NODE_RES level are the same.

This patch changes the locking of NODE_RES such that those locks are
not acquired opportunistically any more. Instead, the mandatory
locks are set to the acquired NODE locks once they are actually
granted.

This fixes issue 622.

Signed-off-by: Thomas Thrainer <thomasth@google.com>
Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>