(kvm) Hotplug a device found only in config.data
A device can be added *only* in config.data by omitting `--hotplug`
option. With this patch we allow this device to be hot-plugged
in the future by issuing `gnt-instance modify --net <index>: --hotplug`
command. HotModDevice() will search the runtime entry of the device
which will raise HoptlugError because it will be not present.
We catch this error and hot-add the device on an available PCI
slot.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>