(2.11) Add genAndRestArguments :: Gen (Map String JSValue)
So that objects using AndRestArguments are available for testing.
As the AndRestArguments are intended for passing through additional
parameters passed on the command line, we restrict them to the
values that are allowed at this position: Strings of ASCII characters
and Integers.

Also, another property guaranteed by the specification is that the
keys in the map are different from all the other fields. To obtain
this property without adding additional complexity to the test
infrastructure, we choose a length for the keys that is longer than
all the valid fields.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>