(2.12) QMP: fetch supported commands on connect
Store the set of supported QMP commands in
QmpConnection.supported_commands. This allows selective error handling
to give the caller a chance to downgrade to HMP when unsupported
commands are encountered.

The `query-commands` QMP command used to implement this, appeared in
QEMU 0.14.0 together with QMP itself, so it is supported by all
instances featuring a QMP monitor.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Signed-off-by: Thomas Thrainer <thomasth@google.com>
Reviewed-by: Thomas Thrainer <thomasth@google.com>