(2.13) Implement GetUserspaceAccessUri for ExtStorage
Allow ExtStorage devices to support userspace access.
The 'attach' script of an ExtStorage provider is now allowed to return
more than one line. The first line will contain as always the block
device path. Each one of the extra lines will contain a URI to be used
for the userspace access by a specific hypervisor. Each URI should be
prefixed with the hypervisor it corresponds to (e.g. kvm:<uri>). The
prefix will be case insensitive.

In order to be backwards compatible, providers that do not return any
extra line in their attach scripts, will be considered to not allow
userspace access for their disks.

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>
Signed-off-by: Thomas Thrainer <thomasth@google.com>
Reviewed-by: Thomas Thrainer <thomasth@google.com>
Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>