qmp: Refactor of add-fd and remove-fd commands
Instead of sending the qmp command along with the SCM_RIGHTS control
message we send just a blank. Then we invoke the corresponding
command (add-fd, remove-fd) with the regular way (qmp.Execute).
Since the qmp connection does not close in between the fds will be
added properly.

AddFd() wrapper will be used for disk hot-add except for the case
where we have userspace access mode. Make it raise HypervisorError
if the command fails for some reason so that any hotplug action that
takes place will abort.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>