hotplug: New _with_qmp decorator
Introduce _with_qmp() decorator that is supposed to keep a shared
qmp connection among hotplug related methods. This is required upon
sending fds with SCM_RIGHTS because qemu cleans up any unused fds if
all monitor connections are closed.

The decorator takes the hypervisor and the instance object as passed
to all hotplug related commands and ensures a valid QmpConnection()
object is found in hypervisor's qmp attribute. Then after the method
return the qmp connection is closed. This way if we have nested
hotplug methods there is no need to change their prototype in order
to pass the QmpConnection object and no need to worry about proper
opening, connecting, and closing.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>