(2.13) netdev: Refactor OpenTap for future use
Make OpenTap() helper method to take a dictionary with the features
that should be enabled or not. Currently these features are: vhost,
vnet_hdr, multiqueue.

In case vhost is enabled, along with the tapfds we return a list of
vhostfds after opening /dev/vhost-net. This regularly is done by the
qemu process if vhost=on is passed in --netdev option, but in case
of hotplug and if the qemu process does not run with root privileges
we must get the fds and pass them via SCM_RIGHTS prior qemu uses
them.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Reviewed-by: Hrvoje Ribicic <riba@google.com>